Fixed conversion warning

+++ kEngine/main.cpp
@@ -103,7 +103,7 @@ int main(int argv, char** args)
     }
 
     // TODO: Create Engine class to manage SDL window, renderer etc.
-    SDL_Window* window = SDL_CreateWindow("kEngine", SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED, screenSize.x, screenSize.y, SDL_WINDOW_OPENGL);
+    SDL_Window* window = SDL_CreateWindow("kEngine", SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED, static_cast<int>(screenSize.x), static_cast<int>(screenSize.y), SDL_WINDOW_OPENGL);
     //SDL_Renderer* renderer = SDL_CreateRenderer(window, -1, SDL_RENDERER_ACCELERATED | SDL_RENDERER_PRESENTVSYNC);
 
     Camera camera(Vector2D(screenSize.x / 2, screenSize.y / 2), 1.0f);
